Update gcc compile flags
+ There was a bug in `config/unix-g++.cmake` that prevented some warning flags from being applied as
  intended.
+ I also added new GCC CXX flags `-Wuseless-cast`, `-Wnull-dereference`, and `-Wnon-virtual-dtor`
+ The `-Wuseless-cast` and reactivation of `-Wold-style-cast` created build warnings that were fixed
  in this PR.  Most of the changes are represented by removing `static_cast<T>` code, although a few
  files required localized pragmas to disable the new warnings.
+ This took a couple of tries to get gcc, llvm and msvc to all compile w/o warnings.

+++ src/RTT_Format_Reader/CellDataIDs.cc
@@ -39,7 +39,7 @@ void CellDataIDs::readData(ifstream &meshfile) {
   int dataIDNum;
   string dummyString;
 
-  for (size_t i = 0; i < static_cast<size_t>(dims.get_ncell_data()); ++i) {
+  for (size_t i = 0; i < dims.get_ncell_data(); ++i) {
     Check(i < names.size() && i < units.size());
     meshfile >> dataIDNum >> names[i] >> units[i];
     Insist(static_cast<size_t>(dataIDNum) == i + 1, "Invalid mesh file: cell data ID out of order");
